+++ src/Command/Arguments.php
@@ -23,7 +23,7 @@ final class Arguments
      * @param Map<string, string> $arguments
      * @param Sequence<string> $pack
      */
-    public function __construct(Map $arguments = null, Sequence $pack = null)
+    public function __construct(?Map $arguments = null, ?Sequence $pack = null)
     {
         $this->arguments = $arguments ?? Map::of();
         $this->pack = $pack ?? Sequence::strings();